+++ nautilus/__init__.py
@@ -5,4 +5,4 @@ from .prior import Prior
 
 __author__ = 'Johannes U. Lange'
 __email__ = 'julange.astro@pm.me'
-__version__ = '0.6.0'
+__version__ = '0.7.0'