+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to nautilus will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Early stopping for neural network training is now default.
+
 ## [0.6.0] - 2023-04-22
 
 ### Changed